% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-template.R
\name{create_minutes}
\alias{create_minutes}
\title{Create minute template}
\usage{
create_minutes(file_name = NULL, ext_name = "minutes")
}
\arguments{
\item{file_name}{The name of the file to save the template shell to.}

\item{ext_name}{The name of the specific extension}
}
\description{
\code{create_minutes} creates a minute template for you to fill in
}
\details{
This is a helper function to create minutes in a consistent way (but it won't
get you any additional hours in the day).
You can use it at any meeting to record the agenda, attendees etc.
}
