% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-template.R
\name{create_minutes_md}
\alias{create_minutes_md}
\title{Create minute template}
\usage{
create_minutes_md(file_name = NULL, ext_name = "minutes-md")
}
\arguments{
\item{file_name}{The name of the file to save the template shell to.}

\item{ext_name}{The name of the specific extension}
}
\description{
\code{create_minutes_md} creates a minute template for you to fill in
}
\details{
This is a helper function to create minutes in a consistent way.
You can use it at any meeting to record the agenda, attendees etc.
}
