% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-template.R
\name{review_release_notif}
\alias{review_release_notif}
\title{Create review release outline document}
\usage{
review_release_notif(file_name = NULL, ext_name = "review-release-notif")
}
\arguments{
\item{file_name}{The name of the file to save the template shell to.}

\item{ext_name}{The name of the specific extension}
}
\description{
\code{review_release_notif} creates a review release template for you to fill in
}
\details{
This is a helper function to create a review release template.
When developing your sap, code, trial report etc, you can use this to produce
a nicely formatted document that captures all the relevant information relevant
to the release notification which signifies a milestone in the products dev.
You can keep this review request as part of your project files and, if
necessary, also transfer a copy to the TMF.
}
