% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-template.R
\name{review_update_notif}
\alias{review_update_notif}
\title{Create review update outline document}
\usage{
review_update_notif(file_name = NULL, ext_name = "review-update-notif")
}
\arguments{
\item{file_name}{The name of the file to save the template shell to.}

\item{ext_name}{The name of the specific extension}
}
\description{
\code{review_update_notif} creates a review update template for you to fill in
}
\details{
This is a helper function to create a review update template.
When developing your sap, code, trial report etc, you can use this to produce
a nicely formatted document that captures all the relevant information relevant
to the updates you committed to making as an outcome of the review. You can
keep this review request as part of your project files and, if necessary,
also transfer a copy to the TMF.
}
